+++ v03_pipeline/lib/tasks/base/base_write.py
