+++ v03_pipeline/lib/tasks/update_lookup_table_with_deleted_families.py
@@ -4,7 +4,7 @@ import luigi
 from v03_pipeline.lib.misc.lookup import (
     remove_family_guids,
 )
-from v03_pipeline.lib.tasks.base.base_update_lookup_table_task import (
+from v03_pipeline.lib.tasks.base.base_update_lookup_table import (
     BaseUpdateLookupTableTask,
 )
 from v03_pipeline.lib.tasks.update_variant_annotations_table_with_deleted_families import (
