+++ v03_pipeline/lib/tasks/update_project_table_with_deleted_families.py
@@ -2,7 +2,7 @@ import hail as hl
 import luigi
 
 from v03_pipeline.lib.misc.family_entries import remove_family_guids
-from v03_pipeline.lib.tasks.base.base_update_project_table_task import (
+from v03_pipeline.lib.tasks.base.base_update_project_table import (
     BaseUpdateProjectTableTask,
 )
