+++ v03_pipeline/lib/tasks/write_relatedness_check_table.py
@@ -7,7 +7,7 @@ from v03_pipeline.lib.paths import (
     relatedness_check_table_path,
     valid_cached_reference_dataset_query_path,
 )
-from v03_pipeline.lib.tasks.base.base_write_task import BaseWriteTask
+from v03_pipeline.lib.tasks.base.base_write import BaseWriteTask
 from v03_pipeline.lib.tasks.files import GCSorLocalTarget, HailTableTask
 from v03_pipeline.lib.tasks.write_imported_callset import WriteImportedCallsetTask
