+++ v03_pipeline/lib/tasks/write_sex_check_table.py
@@ -3,7 +3,7 @@ import luigi
 
 from v03_pipeline.lib.methods.sex_check import call_sex
 from v03_pipeline.lib.paths import sex_check_table_path
-from v03_pipeline.lib.tasks.base.base_write_task import BaseWriteTask
+from v03_pipeline.lib.tasks.base.base_write import BaseWriteTask
 from v03_pipeline.lib.tasks.files import GCSorLocalTarget
 from v03_pipeline.lib.tasks.write_imported_callset import WriteImportedCallsetTask
